Fix for "gitolite_hook" route error
Due to a change in redmine plugin development there was no more default routes
for plugins. Developers now must configure the routes themselves.
This commit provides the configuration of the default route so that the
post-receive hook can work again and redmine will be notificated of changes in
the gitolite repository.
Although it seems to work on my repository it is not guaranteed to work in any
case.